auth-service/requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
python-dotenv
SQLAlchemy
Werkzeug
psycopg2-binary
pika